fix buildroot crosscompile (#1)
* cmake: remove unsafe host include path for libdrm (use pkgconfig instead)


* cmake: add option ENABLE_COMPILE_FLAGS_FOR_TARGET

Add cmake option ENABLE_COMPILE_FLAGS_FOR_TARGET, defaulting to armv8
enabling the previous hard coded compiler options.

Run cmake with -DENABLE_COMPILE_FLAGS_FOR_TARGET=disabled for no extra
compiler options.

Fixes:

  aarch64-buildroot-linux-gnu-g++.br_real: error: unrecognized command-line option ‘-mfpu=neon-fp-armv8’


* libcamera_vid: fix getline call (needs size_t)

Fixes:

.../libcamera-apps-custom/libcamera_vid.cpp:39:26: error: cannot convert ‘unsigned int*’ to ‘size_t*’ {aka ‘long unsigned int*’}
   39 |    getline(&user_string, &len, stdin);
      |                          ^~~~
      |                          |
      |                          unsigned int*


* libcamera_still: fix getline call (needs size_t)

Fixes:

.../build/libcamera-apps-custom/libcamera_still.cpp:152:26: error: cannot convert ‘unsigned int*’ to ‘size_t*’ {aka ‘long unsigned int*’}
  152 |    getline(&user_string, &len, stdin);
      |                          ^~~~
      |                          |
      |                          unsigned int*


* libcamera_app: fix int64_t constants

Fixes:

.../host/aarch64-buildroot-linux-gnu/sysroot/usr/include/libcamera/libcamera/controls.h: In instantiation of ‘void libcamera::ControlList::set(const libcamera::Control<T>&, const std::initializer_list<_Up>&) [with T = libcamera::Span<const long int>; V = long long int]’:
.../libcamera_app.hpp:331:18:   required from ‘void LibcameraApp<OPTIONS>::StartCamera() [with OPTIONS = StillOptions]’
.../libcamera_jpeg.cpp:35:18:   required from here
.../host/aarch64-buildroot-linux-gnu/sysroot/usr/include/libcamera/libcamera/controls.h:399:14: error: no matching function for call to ‘libcamera::ControlValue::set<libcamera::Span<const long int, 18446744073709551615> >(libcamera::Span<const long long int, 18446744073709551615>)’
  399 |   val->set<T>(Span<const typename std::remove_cv_t<V>>{ value.begin(), value.size() });
      |   ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


* cmake: fix linking with exif, jpeg

Fixes:

  [ 66%] Linking CXX executable libcamera-hello
  .../aarch64-buildroot-linux-gnu/bin/ld: libcommon.so: undefined reference to `jpeg_std_error'
  .../aarch64-buildroot-linux-gnu/bin/ld: libcommon.so: undefined reference to `exif_set_slong'

  [ 92%] Linking CXX executable libcamera-vid
  .../aarch64-buildroot-linux-gnu/bin/ld: libcommon.so: undefined reference to `exif_set_slong'


* cmake: add install target


* cmake: no need to link against ipa_rpi.so

Fixes:
	$ libcamera-hello
	libcamera-hello: error while loading shared libraries: ipa_rpi.so: cannot open shared object file: No such file or directory


* circular_output: fix size_t vs. unsigned int narrowing conversion

Fixes:

  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/circular_output.cpp: In member function ‘virtual void CircularOutput::outputBuffer(void*, size_t, int64_t, uint32_t)’:
  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/circular_output.cpp:72:62: warning: narrowing conversion of ‘size’ from ‘size_t {aka long unsigned int}’ to ‘unsigned int’ inside { } [-Wnarrowing]
    Header header = { size, flags & FLAG_KEYFRAME, timestamp_us };
                                                                ^


* egl_preview: fix XVisualInfo initializer

Fixes:

  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/egl_preview.cpp: In member function ‘void EglPreview::makeWindow(const char*)’:
  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/egl_preview.cpp:219:65: sorry, unimplemented: non-trivial designated initializers not supported
    XVisualInfo visTemplate = { .visualid = (long unsigned int)vid };
                                                                   ^


* jpeg: fix long unsigned int vs. size_t mismatch

Introduce a custom jpeg_mem_len_t as the jpeg_mem_dest() signature
changed since jpegsrc.v9d from 'unsigned long * outsize' to
'size_t * outsize'.

Fixes:

  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/jpeg.cpp:270:41: error: invalid conversion from ‘long unsigned int*’ to ‘size_t*’ {aka ‘unsigned int*’} [-fpermissive]
    270 |     jpeg_mem_dest(&cinfo, &jpeg_buffer, &jpeg_len);
        |                                         ^~~~~~~~~
        |                                         |
        |                                         long unsigned int*


* jpeg: fix bool vs. boolean mismatch

Fixes:

  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/mjpeg_encoder.cpp:57:22: error: cannot convert ‘bool’ to ‘boolean’ in assignment
     57 |  cinfo.raw_data_in = true;
        |                      ^~~~


* mjpeg_encoder: fix long unsigned int vs. size_t mismatch

Introduce a custom jpeg_mem_len_t as the jpeg_mem_dest() signature
changed since jpegsrc.v9d from 'unsigned long * outsize' to
'size_t * outsize'.

Fixes:

  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/mjpeg_encoder.cpp:61:44: error: invalid conversion from ‘long unsigned int*’ to ‘size_t*’ {aka ‘unsigned int*’} [-fpermissive]
     61 |     jpeg_mem_dest(&cinfo, &encoded_buffer, &buffer_len);
        |                                            ^~~~~~~~~~~
        |                                            |
        |                                            long unsigned int*


* mjpeg_encoder: fix bool vs. boolean mismatch

Fixes:

  .../libcamera-apps-b1ca8997530f1f0290b291a956de359f2d4ce3e2/mjpeg_encoder.cpp:57:22: error: cannot convert ‘bool’ to ‘boolean’ in assignment
     57 |  cinfo.raw_data_in = true;
        |                      ^~~~
  compilation terminated due to -Wfatal-errors.

+++ mjpeg_encoder.hpp
@@ -55,7 +55,7 @@ private:
 	std::condition_variable encode_cond_var_;
 	std::thread encode_thread_[NUM_ENC_THREADS];
 	void encodeJPEG(struct jpeg_compress_struct &cinfo, EncodeItem &item,
-					uint8_t *&encoded_buffer, long unsigned int &buffer_len);
+					uint8_t *&encoded_buffer, size_t &buffer_len);
 
 	struct OutputItem
 	{
